Move some example plugins to tests.support
